Fix bug that locks weren't being deleted in onbeforeunload handler
Fixes FUN-2077

Browsers have deprecated async AJAX calls, so $.ajax({async: true}) no
longer does anything in recent version of Chrome, Firefox, Safari, or
Edge. Instead, the recommendation is to use the Navigator.sendBeacon API
for http requests that need to be executed after a user leaves a page.

If the browser has support for the sendBeacon method, it is used instead
of the async AJAX call for submit and beforeunload events.

Since sendBeacon only sends POST requests, and the locking API uses HTTP
methods in a REST-like fashion, I've had to enable the ability to use a
GET parameter to call the DELETE handler (specifically, with
?method=delete).