Fix race condition bug when taking over a lock
The change_view locking javascript was making an AJAX call to get a
lock, even if the user had just requested ownership of the lock. It was
then making a subsequent request to forcibly take the lock. If the first
AJAX call succeeded before the second, it would popup an alert telling
the user the article was locked by somebody else.

This change makes it so that it only ever calls
LockingForm.prototype.getLock() or LockingForm.prototype.takeLock() when
initializing, but never both.